requirements.txt
django==4.1.1
djangorestframework==3.13.1
djangorestframework_simplejwt==5.2.0